requirements.txt
torch>=2.0.0
transformers==4.47.0
peft>=0.7.0
trl==0.12.0
accelerate>=0.24.0
datasets>=2.14.0
bitsandbytes>=0.43.1
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart
pydantic>=2.0.0
groq>=0.4.0
pymupdf
scikit-learn
rouge-score
wandb
tqdm
python-dotenv
numpy
